% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityApp_functions.R
\name{add_task}
\alias{add_task}
\title{Add Task}
\usage{
add_task(task, duration, importance, urgency, enjoyability)
}
\arguments{
\item{task}{character, name of the task you want to add}

\item{duration}{integer, duration of the task in minutes}

\item{importance}{integer (1-4), how important the task is}

\item{urgency}{integer(1-4), how urgent the task is}

\item{enjoyability}{integer(1-3), how enjoyable the task is}
}
\value{
the tasks.csv file is updated
}
\description{
Adds a new task with its variables to the tasks.csv file.
}
\examples{
# Add a task to your task file
add_task(task = "Programming", duration = 180, importance = 4, urgency = 3, enjoyability = 2)
}
