% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityApp_functions.R
\name{create_empty_task_csv}
\alias{create_empty_task_csv}
\title{Create an Empty Task File}
\usage{
create_empty_task_csv(filename = "tasks.csv")
}
\arguments{
\item{filename}{string with a default value "tasks.csv"}
}
\value{
csv file is created
}
\description{
Creates an empty csv file with a default name tasks.csv. The file has 5 columns named task, duration, importance, urgency, and enjoyability.
}
\examples{
# Create an empty task csv file in your current working directory
create_empty_task_csv()
}
