% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityApp_functions.R
\name{create_empty_task_df}
\alias{create_empty_task_df}
\title{Create an Empty Task Data Frame}
\usage{
create_empty_task_df()
}
\value{
An empty data frame with columns Task, Duration, Importance, Urgency, and Enjoyability.
}
\description{
Creates an empty task data frame. The data frame has 5 columns named Task, Duration, Importance, Urgency, and Enjoyability. It does not automatically convert strings into factors.
}
\examples{
# Create an empty task data frame 
create_empty_task_df()
}
