% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityApp_functions.R
\name{generate_eisenhower_matrix}
\alias{generate_eisenhower_matrix}
\title{Generate an Eisenhower Matrix}
\usage{
generate_eisenhower_matrix(filename = "tasks.csv")
}
\arguments{
\item{filename}{character, name of the task csv file, default is "tasks.csv"}
}
\value{
the Eisenhower matrix
}
\description{
Creates an Eisenhower matrix based on a file with tasks.
}
\examples{
# Generate an Eisenhower matrix, assuming you have "tasks.csv" file in your working directory
generate_eisenhower_matrix()
}
