% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityApp_functions.R
\name{generate_todolist}
\alias{generate_todolist}
\title{Generate a To-Do List}
\usage{
generate_todolist(
  df,
  start_enjoy = FALSE,
  start_short = TRUE,
  short_time = 5,
  start_time = "9:00",
  available_time = 480,
  show_intervals = TRUE
)
}
\arguments{
\item{df}{name of the task data frame}

\item{start_enjoy}{logical, whether user prefers to start with most or least enjoyable tasks that have equal urgency and importance.}

\item{start_short}{logical, whether user wants to start with short tasks (e.g. tasks that take less than 5 mins) regardless of their urgency, importance, and enjoyability.}

\item{short_time}{numeric, what task duration user considers to be short, this parameter is relevant if start_short is TRUE, default value is 5 minutes.}

\item{start_time}{character in a format "H:M", represents the time when the user wants to start working on their tasks.}

\item{available_time}{numeric, how much available time the user has (in minutes).}

\item{show_intervals}{logical, whether user wants to have time intervals next to every task}
}
\value{
the to-do list data frame with time intervals and tasks
}
\description{
Generates a to-do list based on a given task data frame. It uses the order_task_df function to first order the tasks based on urgency, importance and user preferences. Next, it assigns time intervals to each task and returns a data frame containing a to-do list with intervals and tasks.
}
\examples{
# Generate a to-do list based on this example task data frame
example_df = data.frame(Task = c("Meeting preparation", "Programming", "Watch lecture recording", "Make a dentist appointment for August", "Update my CV"),Duration = c(30, 60, 75, 3, 30),Importance = c(4, 3, 3, 2, 4),Urgency = c(4, 2, 3, 4, 4),Enjoyability = c(1, 3, 2, 1, 2),stringsAsFactors = FALSE)
generate_todolist(example_df)
}
