% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityApp_functions.R
\name{get_productivity_plot}
\alias{get_productivity_plot}
\title{Generate a Productivity Plot}
\usage{
get_productivity_plot(df)
}
\arguments{
\item{df}{name of the data frame with tasks and estimated and actual duration of each task}
}
\value{
The productivity plot with estimated and actual duration of individual tasks.
}
\description{
Generates a productivity plot with estimated and actual duration of individual tasks. Time is shown on the x-axis and tasks on the y-axis.
}
\examples{
# Generate a productivity plot based on an example data frame
example_df <- data.frame(Task = c("Meeting preparation", "Programming", "Watch lecture recording"), Estimated_duration = c(30, 120, 90), Actual_duration = c(41, 200, 85))
get_productivity_plot(example_df)
}
