% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityApp_functions.R
\name{order_task_df}
\alias{order_task_df}
\title{Order Task Data Frame}
\usage{
order_task_df(df, start_enjoy = FALSE, start_short = TRUE, short_time = 5)
}
\arguments{
\item{df}{name of the task data frame}

\item{start_enjoy}{logical, whether user prefers to start with most or least enjoyable tasks that have equal urgency and importance.}

\item{start_short}{logical, whether user wants to start with short tasks (e.g. tasks that take less than 5 mins) regardless of their urgency, importance, and enjoyability.}

\item{short_time}{numeric, what task duration user considers to be short, this parameter is relevant if start_short is TRUE, default value is 5 minutes.}
}
\value{
the ordered task data frame with the following columns:
\itemize{
\item \strong{Task}: the names of individual tasks
\item \strong{Duration}: the estimated duration of tasks
\item \strong{Importance}: how important the task is on a scale 1-4
\item \strong{Urgency}: how urgent the task is on a scale 1-4
\item \strong{Enjoyability}: how enjoyable the task is on a scale 1-3
}
}
\description{
Orders the task data frame based on urgency, importance, and user preferences. Urgency takes precedence before importance and importance takes precedence before enjoyability.
}
\examples{
# Order an example task data frame
example_df = data.frame(Task = c("Meeting preparation", "Programming", "Watch lecture recording", "Make a dentist appointment for August", "Update my CV"),Duration = c(30, 60, 75, 3, 30),Importance = c(4, 3, 3, 2, 4),Urgency = c(4, 2, 3, 4, 4),Enjoyability = c(1, 3, 2, 1, 2),stringsAsFactors = FALSE)
order_task_df(example_df)
}
