% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityApp_functions.R
\name{tasks_given_availability}
\alias{tasks_given_availability}
\title{Check Tasks Given Availability}
\usage{
tasks_given_availability(df, start_time, end_time)
}
\arguments{
\item{df}{name of the ordered task data frame}

\item{start_time}{start time in a format "\%H:\%M", when user wants to start working on their to-do list}

\item{end_time}{end time in a format "\%H:\%M", until when user wants to work on their to-do list}
}
\value{
df and message:
#' \itemize{
\item \strong{Data Frame}: a data frame with tasks that can be scheduled based on user's availability.
\item \strong{Message}: a message about whether some tasks couldn't be scheduled - either saying "All tasks were successfully scheduled!" or "The following tasks were not scheduled because there was not enough time available:" followed by the names of tasks that could not be scheduled.
}
}
\description{
Generates a data frame and a message about unscheduled tasks for a to-do list based on user's time availability.
}
\examples{
# Generate a data frame and a message about unscheduled tasks
example_df <- data.frame(Task = c("Meeting preparation", "Programming", "Watch lecture recording", "Make a dentist appointment for August", "Update my CV"),Duration = c(30, 60, 75, 3, 30),Importance = c(4, 3, 3, 2, 4),Urgency = c(4, 2, 3, 4, 4),Enjoyability = c(1, 3, 2, 1, 2),stringsAsFactors = FALSE)
example_ordered_df <- order_task_df(example_df)
tasks_given_availability(example_ordered_df)
}
