% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityApp_functions.R
\name{generate_eisenhower_matrix}
\alias{generate_eisenhower_matrix}
\title{Generate an Eisenhower Matrix}
\usage{
generate_eisenhower_matrix(df)
}
\arguments{
\item{df}{name of the task data frame}
}
\value{
the Eisenhower matrix
}
\description{
Creates an Eisenhower matrix based on a data frame with tasks. Eisenhower matrix shows tasks' urgency on the x-axis and importance on the y-axis. It provides a good way to visualize tasks that should be prioritized.
}
\examples{
# Generate an Eisenhower matrix based on an example data frame
example_df = data.frame(Task = c("Meeting preparation", "Programming", "Watch lecture recording"),Duration = c(30, 60, 75),Importance = c(4, 3, 3),Urgency = c(4, 2, 3),Enjoyability = c(1, 3, 2),stringsAsFactors = FALSE)
generate_eisenhower_matrix(df = example_df)
}
